+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased (3.3.0)
 
 ### Added
+- Added the Min Levels field setting (added `benf\neo\Field::$minLevels`)
 - Added the Min Blocks block type setting (added `benf\neo\models\BlockType::$minBlocks` and added the `minBlocks` column to the `neoblocktypes` table)
 - Added the Min Sibling Blocks of This Type block type setting (added `benf\neo\models\BlockType::$minSiblingBlocks` and added the `minSiblingBlocks` column to the `neoblocktypes` table)
 - Added the Min Child Blocks block type setting (added `benf\neo\models\BlockType::$minChildBlocks` and added the `minChildBlocks` column to the `neoblocktypes` table)